+++ apps/meteor/tests/data/apps/apps-data.ts
@@ -1,4 +1,3 @@
 export const APP_URL = 'https://github.com/RocketChat/Apps.RocketChat.Tester/blob/master/dist/appsrocketchattester_0.0.5.zip?raw=true';
 export const APP_NAME = 'Apps.RocketChat.Tester';
-export const APP_USERNAME = 'appsrocketchattester.bot';
 export const apps = (path = '') => `/api/apps${path}` as const;
