+++ apps/meteor/tests/data/chat.helper.ts
@@ -29,16 +29,6 @@ export const sendSimpleMessage = ({
 	return request.post(api('chat.sendMessage')).set(credentials).send({ message });
 };
 
-export const pinMessage = ({ msgId }: { msgId: IMessage['_id'] }) => {
-	if (!msgId) {
-		throw new Error('"msgId" is required in "pinMessage" test helper');
-	}
-
-	return request.post(api('chat.pinMessage')).set(credentials).send({
-		messageId: msgId,
-	});
-};
-
 export const deleteMessage = ({ roomId, msgId }: { roomId: IRoom['_id']; msgId: IMessage['_id'] }) => {
 	if (!roomId) {
 		throw new Error('"roomId" is required in "deleteMessage" test helper');
