+++ apps/meteor/tests/data/constants.ts
@@ -1,3 +1,1 @@
 export const CI_MAX_ROOMS_PER_GUEST = 10;
-export const MAX_BIO_LENGTH = 260;
-export const MAX_NICKNAME_LENGTH = 120;
