+++ apps/meteor/tests/data/interactions.ts
@@ -1,6 +1,1 @@
-export const targetUser = 'rocket.cat';
 export const imgURL = './public/images/logo/1024x1024.png';
-export const lstURL = './tests/e2e/fixtures/files/lst-test.lst';
-export const drawioURL = './tests/e2e/fixtures/files/diagram.drawio';
-export const svgLogoURL = './public/images/logo/logo.svg';
-export const svgLogoFileName = 'logo.svg';
