--- apps/meteor/tests/data/role.ts
@@ -1,5 +1,0 @@
-export const roleNameUsers = `role-name-test-users-${Date.now()}`;
-export const roleNameSubscriptions = `role-name-test-subscriptions-${Date.now()}`;
-export const roleScopeUsers = 'Users';
-export const roleScopeSubscriptions = 'Subscriptions';
-export const roleDescription = `role-description-test-${Date.now()}`;
